move dynamic costs def
